requirements.txt
numpy
PyYAML
tqdm
setuptools
black>24.3.0
ufmt==2.0.0
usort==1.0.4
imageio[pyav]
imageio
imageio-ffmpeg
